feat: remove beat sequnce to audio engine
